refactor(podman): extract WinVersionCheck to a dedicated file
